modified 3-isalpha.c and 4-isalpha.c

+++ 0x02-functions_nested_loops/3-islower.c
@@ -7,5 +7,5 @@
  */
 int _islower(char c)
 {
-return c >= 'a' && c <= 'z';
+return (int) c >= 'a' && c <= 'z';
 }
